backend/requirements.txt
fastapi
uvicorn
paramiko
pywinrm
python-multipart
